+++ app/src/androidTest/java/com/the_tinkering/wk/test/DatabaseMigrationTest.java
@@ -50,6 +50,7 @@ import static com.the_tinkering.wk.db.AppDatabase.MIGRATION_63_64;
 import static com.the_tinkering.wk.db.AppDatabase.MIGRATION_64_65;
 import static com.the_tinkering.wk.db.AppDatabase.MIGRATION_65_66;
 import static com.the_tinkering.wk.db.AppDatabase.MIGRATION_66_67;
+import static com.the_tinkering.wk.db.AppDatabase.MIGRATION_67_68;
 import static java.util.Objects.requireNonNull;
 import static org.junit.Assert.assertEquals;
 
@@ -59,7 +60,7 @@ import static org.junit.Assert.assertEquals;
 @SuppressWarnings({"JavaDoc", "resource"})
 @SmallTest
 public final class DatabaseMigrationTest {
-    private static final int LATEST_VERSION = 67;
+    private static final int LATEST_VERSION = 68;
     private static final String DATABASE_NAME_TEST = "wanikani-test";
 
     @Rule
@@ -91,7 +92,8 @@ public final class DatabaseMigrationTest {
                         MIGRATION_63_64,
                         MIGRATION_64_65,
                         MIGRATION_65_66,
-                        MIGRATION_66_67)
+                        MIGRATION_66_67,
+                        MIGRATION_67_68)
                 .build();
         testHelper.closeWhenFinished(database);
         return database;
@@ -106,7 +108,7 @@ public final class DatabaseMigrationTest {
         testHelper.runMigrationsAndValidate(DATABASE_NAME_TEST, LATEST_VERSION, true, MIGRATION_48_49, MIGRATION_49_50, MIGRATION_50_51,
                 MIGRATION_51_52, MIGRATION_52_53, MIGRATION_53_54, MIGRATION_54_55, MIGRATION_55_56, MIGRATION_56_57, MIGRATION_57_58,
                 MIGRATION_58_59, MIGRATION_59_60, MIGRATION_60_61, MIGRATION_61_62, MIGRATION_62_63, MIGRATION_63_64, MIGRATION_64_65,
-                MIGRATION_65_66, MIGRATION_66_67);
+                MIGRATION_65_66, MIGRATION_66_67, MIGRATION_67_68);
         assertEquals(LATEST_VERSION, getMigratedRoomDatabase().getOpenHelper().getReadableDatabase().getVersion());
     }
 
@@ -119,7 +121,7 @@ public final class DatabaseMigrationTest {
         testHelper.runMigrationsAndValidate(DATABASE_NAME_TEST, LATEST_VERSION, true, MIGRATION_49_50, MIGRATION_50_51, MIGRATION_51_52,
                 MIGRATION_52_53, MIGRATION_53_54, MIGRATION_54_55, MIGRATION_55_56, MIGRATION_56_57, MIGRATION_57_58, MIGRATION_58_59,
                 MIGRATION_59_60, MIGRATION_60_61, MIGRATION_61_62, MIGRATION_62_63, MIGRATION_63_64, MIGRATION_64_65, MIGRATION_65_66,
-                MIGRATION_66_67);
+                MIGRATION_66_67, MIGRATION_67_68);
         assertEquals(LATEST_VERSION, getMigratedRoomDatabase().getOpenHelper().getReadableDatabase().getVersion());
     }
 
@@ -131,7 +133,8 @@ public final class DatabaseMigrationTest {
 
         testHelper.runMigrationsAndValidate(DATABASE_NAME_TEST, LATEST_VERSION, true, MIGRATION_50_51, MIGRATION_51_52, MIGRATION_52_53,
                 MIGRATION_53_54, MIGRATION_54_55, MIGRATION_55_56, MIGRATION_56_57, MIGRATION_57_58, MIGRATION_58_59, MIGRATION_59_60,
-                MIGRATION_60_61, MIGRATION_61_62, MIGRATION_62_63, MIGRATION_63_64, MIGRATION_64_65, MIGRATION_65_66, MIGRATION_66_67);
+                MIGRATION_60_61, MIGRATION_61_62, MIGRATION_62_63, MIGRATION_63_64, MIGRATION_64_65, MIGRATION_65_66, MIGRATION_66_67,
+                MIGRATION_67_68);
         assertEquals(LATEST_VERSION, getMigratedRoomDatabase().getOpenHelper().getReadableDatabase().getVersion());
     }
 
@@ -143,7 +146,7 @@ public final class DatabaseMigrationTest {
 
         testHelper.runMigrationsAndValidate(DATABASE_NAME_TEST, LATEST_VERSION, true, MIGRATION_51_52, MIGRATION_52_53, MIGRATION_53_54,
                 MIGRATION_54_55, MIGRATION_55_56, MIGRATION_56_57, MIGRATION_57_58, MIGRATION_58_59, MIGRATION_59_60, MIGRATION_60_61,
-                MIGRATION_61_62, MIGRATION_62_63, MIGRATION_63_64, MIGRATION_64_65, MIGRATION_65_66, MIGRATION_66_67);
+                MIGRATION_61_62, MIGRATION_62_63, MIGRATION_63_64, MIGRATION_64_65, MIGRATION_65_66, MIGRATION_66_67, MIGRATION_67_68);
         assertEquals(LATEST_VERSION, getMigratedRoomDatabase().getOpenHelper().getReadableDatabase().getVersion());
     }
 
@@ -155,7 +158,7 @@ public final class DatabaseMigrationTest {
 
         testHelper.runMigrationsAndValidate(DATABASE_NAME_TEST, LATEST_VERSION, true, MIGRATION_52_53, MIGRATION_53_54, MIGRATION_54_55,
                 MIGRATION_55_56, MIGRATION_56_57, MIGRATION_57_58, MIGRATION_58_59, MIGRATION_59_60, MIGRATION_60_61, MIGRATION_61_62,
-                MIGRATION_62_63, MIGRATION_63_64, MIGRATION_64_65, MIGRATION_65_66, MIGRATION_66_67);
+                MIGRATION_62_63, MIGRATION_63_64, MIGRATION_64_65, MIGRATION_65_66, MIGRATION_66_67, MIGRATION_67_68);
         assertEquals(LATEST_VERSION, getMigratedRoomDatabase().getOpenHelper().getReadableDatabase().getVersion());
     }
 
@@ -167,7 +170,7 @@ public final class DatabaseMigrationTest {
 
         testHelper.runMigrationsAndValidate(DATABASE_NAME_TEST, LATEST_VERSION, true, MIGRATION_53_54, MIGRATION_54_55, MIGRATION_55_56,
                 MIGRATION_56_57, MIGRATION_57_58, MIGRATION_58_59, MIGRATION_59_60, MIGRATION_60_61, MIGRATION_61_62, MIGRATION_62_63,
-                MIGRATION_63_64, MIGRATION_64_65, MIGRATION_65_66, MIGRATION_66_67);
+                MIGRATION_63_64, MIGRATION_64_65, MIGRATION_65_66, MIGRATION_66_67, MIGRATION_67_68);
         assertEquals(LATEST_VERSION, getMigratedRoomDatabase().getOpenHelper().getReadableDatabase().getVersion());
     }
 
@@ -179,7 +182,7 @@ public final class DatabaseMigrationTest {
 
         testHelper.runMigrationsAndValidate(DATABASE_NAME_TEST, LATEST_VERSION, true, MIGRATION_54_55, MIGRATION_55_56, MIGRATION_56_57,
                 MIGRATION_57_58, MIGRATION_58_59, MIGRATION_59_60, MIGRATION_60_61, MIGRATION_61_62, MIGRATION_62_63, MIGRATION_63_64,
-                MIGRATION_64_65, MIGRATION_65_66, MIGRATION_66_67);
+                MIGRATION_64_65, MIGRATION_65_66, MIGRATION_66_67, MIGRATION_67_68);
         assertEquals(LATEST_VERSION, getMigratedRoomDatabase().getOpenHelper().getReadableDatabase().getVersion());
     }
 
@@ -191,7 +194,7 @@ public final class DatabaseMigrationTest {
 
         testHelper.runMigrationsAndValidate(DATABASE_NAME_TEST, LATEST_VERSION, true, MIGRATION_55_56, MIGRATION_56_57, MIGRATION_57_58,
                 MIGRATION_58_59, MIGRATION_59_60, MIGRATION_60_61, MIGRATION_61_62, MIGRATION_62_63, MIGRATION_63_64, MIGRATION_64_65,
-                MIGRATION_65_66, MIGRATION_66_67);
+                MIGRATION_65_66, MIGRATION_66_67, MIGRATION_67_68);
         assertEquals(LATEST_VERSION, getMigratedRoomDatabase().getOpenHelper().getReadableDatabase().getVersion());
     }
 
@@ -203,7 +206,7 @@ public final class DatabaseMigrationTest {
 
         testHelper.runMigrationsAndValidate(DATABASE_NAME_TEST, LATEST_VERSION, true, MIGRATION_56_57, MIGRATION_57_58, MIGRATION_58_59,
                 MIGRATION_59_60, MIGRATION_60_61, MIGRATION_61_62, MIGRATION_62_63, MIGRATION_63_64, MIGRATION_64_65, MIGRATION_65_66,
-                MIGRATION_66_67);
+                MIGRATION_66_67, MIGRATION_67_68);
         assertEquals(LATEST_VERSION, getMigratedRoomDatabase().getOpenHelper().getReadableDatabase().getVersion());
     }
 
@@ -214,7 +217,8 @@ public final class DatabaseMigrationTest {
         db.close();
 
         testHelper.runMigrationsAndValidate(DATABASE_NAME_TEST, LATEST_VERSION, true, MIGRATION_57_58, MIGRATION_58_59, MIGRATION_59_60,
-                MIGRATION_60_61, MIGRATION_61_62, MIGRATION_62_63, MIGRATION_63_64, MIGRATION_64_65, MIGRATION_65_66, MIGRATION_66_67);
+                MIGRATION_60_61, MIGRATION_61_62, MIGRATION_62_63, MIGRATION_63_64, MIGRATION_64_65, MIGRATION_65_66, MIGRATION_66_67,
+                MIGRATION_67_68);
         assertEquals(LATEST_VERSION, getMigratedRoomDatabase().getOpenHelper().getReadableDatabase().getVersion());
     }
 
@@ -225,7 +229,7 @@ public final class DatabaseMigrationTest {
         db.close();
 
         testHelper.runMigrationsAndValidate(DATABASE_NAME_TEST, LATEST_VERSION, true, MIGRATION_58_59, MIGRATION_59_60, MIGRATION_60_61,
-                MIGRATION_61_62, MIGRATION_62_63, MIGRATION_63_64, MIGRATION_64_65, MIGRATION_65_66, MIGRATION_66_67);
+                MIGRATION_61_62, MIGRATION_62_63, MIGRATION_63_64, MIGRATION_64_65, MIGRATION_65_66, MIGRATION_66_67, MIGRATION_67_68);
         assertEquals(LATEST_VERSION, getMigratedRoomDatabase().getOpenHelper().getReadableDatabase().getVersion());
     }
 
@@ -236,7 +240,7 @@ public final class DatabaseMigrationTest {
         db.close();
 
         testHelper.runMigrationsAndValidate(DATABASE_NAME_TEST, LATEST_VERSION, true, MIGRATION_59_60, MIGRATION_60_61, MIGRATION_61_62,
-                MIGRATION_62_63, MIGRATION_63_64, MIGRATION_64_65, MIGRATION_65_66, MIGRATION_66_67);
+                MIGRATION_62_63, MIGRATION_63_64, MIGRATION_64_65, MIGRATION_65_66, MIGRATION_66_67, MIGRATION_67_68);
         assertEquals(LATEST_VERSION, getMigratedRoomDatabase().getOpenHelper().getReadableDatabase().getVersion());
     }
 
@@ -247,7 +251,7 @@ public final class DatabaseMigrationTest {
         db.close();
 
         testHelper.runMigrationsAndValidate(DATABASE_NAME_TEST, LATEST_VERSION, true, MIGRATION_60_61, MIGRATION_61_62, MIGRATION_62_63,
-                MIGRATION_63_64, MIGRATION_64_65, MIGRATION_65_66, MIGRATION_66_67);
+                MIGRATION_63_64, MIGRATION_64_65, MIGRATION_65_66, MIGRATION_66_67, MIGRATION_67_68);
         assertEquals(LATEST_VERSION, getMigratedRoomDatabase().getOpenHelper().getReadableDatabase().getVersion());
     }
 
@@ -258,7 +262,7 @@ public final class DatabaseMigrationTest {
         db.close();
 
         testHelper.runMigrationsAndValidate(DATABASE_NAME_TEST, LATEST_VERSION, true, MIGRATION_61_62, MIGRATION_62_63, MIGRATION_63_64,
-                MIGRATION_64_65, MIGRATION_65_66, MIGRATION_66_67);
+                MIGRATION_64_65, MIGRATION_65_66, MIGRATION_66_67, MIGRATION_67_68);
         assertEquals(LATEST_VERSION, getMigratedRoomDatabase().getOpenHelper().getReadableDatabase().getVersion());
     }
 
@@ -269,7 +273,7 @@ public final class DatabaseMigrationTest {
         db.close();
 
         testHelper.runMigrationsAndValidate(DATABASE_NAME_TEST, LATEST_VERSION, true, MIGRATION_62_63, MIGRATION_63_64, MIGRATION_64_65,
-                MIGRATION_65_66, MIGRATION_66_67);
+                MIGRATION_65_66, MIGRATION_66_67, MIGRATION_67_68);
         assertEquals(LATEST_VERSION, getMigratedRoomDatabase().getOpenHelper().getReadableDatabase().getVersion());
     }
 
@@ -280,7 +284,7 @@ public final class DatabaseMigrationTest {
         db.close();
 
         testHelper.runMigrationsAndValidate(DATABASE_NAME_TEST, LATEST_VERSION, true, MIGRATION_63_64, MIGRATION_64_65, MIGRATION_65_66,
-                MIGRATION_66_67);
+                MIGRATION_66_67, MIGRATION_67_68);
         assertEquals(LATEST_VERSION, getMigratedRoomDatabase().getOpenHelper().getReadableDatabase().getVersion());
     }
 
@@ -290,7 +294,8 @@ public final class DatabaseMigrationTest {
         assertEquals(64, db.getVersion());
         db.close();
 
-        testHelper.runMigrationsAndValidate(DATABASE_NAME_TEST, LATEST_VERSION, true, MIGRATION_64_65, MIGRATION_65_66, MIGRATION_66_67);
+        testHelper.runMigrationsAndValidate(DATABASE_NAME_TEST, LATEST_VERSION, true, MIGRATION_64_65, MIGRATION_65_66, MIGRATION_66_67,
+                MIGRATION_67_68);
         assertEquals(LATEST_VERSION, getMigratedRoomDatabase().getOpenHelper().getReadableDatabase().getVersion());
     }
 
@@ -300,7 +305,7 @@ public final class DatabaseMigrationTest {
         assertEquals(65, db.getVersion());
         db.close();
 
-        testHelper.runMigrationsAndValidate(DATABASE_NAME_TEST, LATEST_VERSION, true, MIGRATION_65_66, MIGRATION_66_67);
+        testHelper.runMigrationsAndValidate(DATABASE_NAME_TEST, LATEST_VERSION, true, MIGRATION_65_66, MIGRATION_66_67, MIGRATION_67_68);
         assertEquals(LATEST_VERSION, getMigratedRoomDatabase().getOpenHelper().getReadableDatabase().getVersion());
     }
 
@@ -310,7 +315,17 @@ public final class DatabaseMigrationTest {
         assertEquals(66, db.getVersion());
         db.close();
 
-        testHelper.runMigrationsAndValidate(DATABASE_NAME_TEST, LATEST_VERSION, true, MIGRATION_66_67);
+        testHelper.runMigrationsAndValidate(DATABASE_NAME_TEST, LATEST_VERSION, true, MIGRATION_66_67, MIGRATION_67_68);
+        assertEquals(LATEST_VERSION, getMigratedRoomDatabase().getOpenHelper().getReadableDatabase().getVersion());
+    }
+
+    @Test
+    public void testMigration_67() throws IOException {
+        final SupportSQLiteDatabase db = testHelper.createDatabase(DATABASE_NAME_TEST, 67);
+        assertEquals(67, db.getVersion());
+        db.close();
+
+        testHelper.runMigrationsAndValidate(DATABASE_NAME_TEST, LATEST_VERSION, true, MIGRATION_67_68);
         assertEquals(LATEST_VERSION, getMigratedRoomDatabase().getOpenHelper().getReadableDatabase().getVersion());
     }
 }